refactor: change notificator to notifier
