+++ pkg/notifier/slack_notifier.go
@@ -1,13 +1,13 @@
-package notificator
+package notifier
 
 import (
 	"github.com/dondakeshimo/todo-cli/pkg/scheduler"
 )
 
-type SlackNotificator struct {
+type SlackNotifier struct {
 	Scheduler scheduler.Scheduler
 }
 
-func (sn *SlackNotificator) Push (r *Request) error {
+func (sn *SlackNotifier) Push (r *Request) error {
 	return nil
 }
